package/rpi-eeprom: New package for RPi bootloader tools
This package adds the host and OS tools for configuring
the bootloader and building/signing tools for signed boot.

Normally, the host tools would be invoked from a post-build
script to sign the boot.img file. The device tools would be
invoked from a custom on-device software update process e.g.
using rpi-eeprom-digest to verify the signature of a boot.img
file against a public key retrieved from the EEPROM.

The current version does not download the EEPROM images
because the bootloader EEPROM is not normally updated via
a buildroot image. However, future versions might provide
the facility to generate both a buildroot initramfs and
accompanying signed EEPROM image e.g for network install.

Signed-off-by: Tim Gover <tim.gover@raspberrypi.com>